replace mui tabs with our adaptation using an array of objects describing each tab + turn each component into its own file
